requirements.txt
streamlit==1.23.1
pdfminer.six
transformers
torch
